scripts/requirements.txt
pandas
numpy
supabase
python-dotenv 